Textures

+++ source/cpp/core/Context.h
@@ -29,6 +29,7 @@
 #include <cfloat>
 #include <vector>
 #include "./../sgl.h"
+#include "./types.h"
 
 /*
  * Side-notes:
@@ -112,7 +113,7 @@ struct Context
 	Vertex emissiveAtt; //< attenuation factor for this emissive light
 
 
-	Context(uint_fast16_t width, uint_fast16_t height)
+	Context(uint16 width, uint16 height)
 	{
 		storage        = Chunk();
 		storage.w      = width;
@@ -157,9 +158,7 @@ struct Context
 	{ scene.getCurrentNode()->addVertex(create(x, y, 0.0f, 1.0f)); }
 
 	inline void setVertex3f(float x, float y, float z)
-	{
-		scene.getCurrentNode()->addVertex(create(x, y, z, 1.0f));
-	}
+	{ scene.getCurrentNode()->addVertex(create(x, y, z, 1.0f)); }
 
 	inline void setVertex4f(float x, float y, float z, float w)
 	{ scene.getCurrentNode()->addVertex(create(x, y, z, w)); }
@@ -264,7 +263,7 @@ struct Context
 		float sa			= sinf(alpha)			;
 		float ca			= cosf(alpha)			;
 
-		int_fast32_t RR     = Helper::round(N)+offset;
+		uint16 RR     = Helper::round(N)+offset;
 
 		switch( drawType )
 		{
@@ -293,7 +292,7 @@ struct Context
 			}break;
 		}
 
-		for (int_fast32_t i = offset; i <= RR; i++)
+		for( uint16 i = offset; i <= RR; i++)
 		{
 			x2 = ca * x1 - sa * y1;
 			y2 = sa * x1 + ca * y1;
@@ -485,7 +484,7 @@ struct Context
 
 	inline void clearDepthBuffer()
 	{
-	    for(uint_fast32_t i = 0; i < storage.w_h; ++i)
+	    for( uint32 i = 0; i<storage.w_h; ++i )
 	        storage.depth[i] = std::numeric_limits<float>::max();;
 	}
 
@@ -524,10 +523,10 @@ struct Context
 		*((__color*) (storage.clear+7))	= *((__color*) &c);
 		*((__color*) (storage.clear+8))	= *((__color*) &c);
 
-		uint_fast32_t  l  =  8					;
-		int_fast8_t    s  =  sizeof(__color)	;
+		uint16   l  =  8				;
+		uint8    s  =  sizeof(__color)	;
 
-		for(uint_fast32_t offset=l ; offset < storage.w_h; offset <<= 1)
+		for( uint32 offset=l ; offset < storage.w_h; offset <<= 1 )
 		{
 			memcpy(&storage.clear[offset], &storage.clear[l], s);
 			l = offset;
